Fix: HTTP Requests through Chromium
animepahe.com got a CloudFlare DDoS protection. Due to that implementation, the old request system doesn't work anymore, so this is a workaround. The poster doesn't work properly yet so it is disabled, I did some progress nonetheless. The whole thing is more unstable and slower than it was before this event occurred.
Additionally I cleaned the interface, no that much but it's ok.